+++ bpm-engine-impl/src/main/java/io/takari/bpm/AbstractEngine.java
@@ -11,7 +11,9 @@ import io.takari.bpm.event.EventPersistenceManager;
 import io.takari.bpm.lock.LockManager;
 import io.takari.bpm.persistence.PersistenceManager;
 import io.takari.bpm.planner.Planner;
+import io.takari.bpm.resource.ResourceResolver;
 import io.takari.bpm.state.*;
+import io.takari.bpm.task.UserTaskHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;