+++ bpm-engine-impl/src/main/java/io/takari/bpm/DefaultExecutor.java
@@ -6,7 +6,9 @@ import io.takari.bpm.el.ExpressionManager;
 import io.takari.bpm.event.EventPersistenceManager;
 import io.takari.bpm.persistence.PersistenceManager;
 import io.takari.bpm.reducers.*;
+import io.takari.bpm.resource.ResourceResolver;
 import io.takari.bpm.state.ProcessInstance;
+import io.takari.bpm.task.UserTaskHandler;
 
 import java.util.List;
 import java.util.concurrent.ExecutorService;
@@ -15,10 +17,16 @@ public class DefaultExecutor implements Executor {
 
     private final Reducer reducer;
 
-    public DefaultExecutor(Configuration cfg, ExpressionManager expressionManager, ExecutorService executor,
-                           ExecutionInterceptorHolder interceptors, IndexedProcessDefinitionProvider definitionProvider,
-                           UuidGenerator uuidGenerator, EventPersistenceManager eventManager,
-                           PersistenceManager persistenceManager) {
+    public DefaultExecutor(Configuration cfg,
+                           ExpressionManager expressionManager,
+                           ExecutorService executor,
+                           ExecutionInterceptorHolder interceptors,
+                           IndexedProcessDefinitionProvider definitionProvider,
+                           UuidGenerator uuidGenerator,
+                           EventPersistenceManager eventManager,
+                           PersistenceManager persistenceManager,
+                           UserTaskHandler userTaskHandler,
+                           ResourceResolver resourceResolver) {
 
         this.reducer = new CombiningReducer(
                 new ForkReducer(expressionManager),
@@ -27,7 +35,7 @@ public class DefaultExecutor implements Executor {
                 new FlowsReducer(),
                 new VariablesReducer(expressionManager),
                 new RaiseErrorReducer(expressionManager),
-                new ExpressionsReducer(expressionManager, executor),
+                new ExpressionsReducer(cfg, expressionManager, executor),
                 new InterceptorEventsReducer(interceptors),
                 new CallActivityReducer(definitionProvider, cfg),
                 new EventsReducer(uuidGenerator, expressionManager, eventManager),
@@ -36,7 +44,9 @@ public class DefaultExecutor implements Executor {
                 new ActivationsReducer(interceptors),
                 new FlowListenerReducer(expressionManager),
                 new ScopeReducer(uuidGenerator),
-                new EventGatewayReducer());
+                new EventGatewayReducer(),
+                new UserTaskReducer(userTaskHandler),
+                new ScriptReducer(resourceResolver, expressionManager));
     }
 
     @Override