+++ bpm-engine-impl/src/main/java/io/takari/bpm/reducers/ExpressionsReducer.java
@@ -1,5 +1,6 @@
 package io.takari.bpm.reducers;
 
+import io.takari.bpm.Configuration;
 import io.takari.bpm.actions.Action;
 import io.takari.bpm.actions.EvalExpressionAction;
 import io.takari.bpm.actions.SetVariableAction;
@@ -14,6 +15,7 @@ import io.takari.bpm.context.ExecutionContextImpl;
 import io.takari.bpm.el.ExpressionManager;
 import io.takari.bpm.model.ExpressionType;
 import io.takari.bpm.model.VariableMapping;
+import io.takari.bpm.model.ServiceTask;
 import io.takari.bpm.state.BpmnErrorHelper;
 import io.takari.bpm.state.ProcessInstance;
 import io.takari.bpm.state.Variables;
@@ -34,10 +36,12 @@ public class ExpressionsReducer implements Reducer {
 
     private static final Logger log = LoggerFactory.getLogger(ExpressionsReducer.class);
 
+    private final Configuration cfg;
     private final ExpressionManager expressionManager;
     private final ExecutorService executor;
 
-    public ExpressionsReducer(ExpressionManager expressionManager, ExecutorService executor) {
+    public ExpressionsReducer(Configuration cfg, ExpressionManager expressionManager, ExecutorService executor) {
+        this.cfg = cfg;
         this.expressionManager = expressionManager;
         this.executor = executor;
     }
@@ -50,10 +54,11 @@ public class ExpressionsReducer implements Reducer {
 
         final EvalExpressionAction a = (EvalExpressionAction) action;
 
-        Variables vars = applyInVariables(state.getVariables(), a.getIn());
+        Variables vars = VariablesHelper.applyInVariables(expressionManager, state.getVariables(), a.getIn());
         final ExecutionContextImpl ctx = new ExecutionContextImpl(expressionManager, vars);
 
-        Callable<Command> fn = new DelegateFn(expressionManager, ctx, a.getType(), a.getExpression(), a.getDefaultCommand());
+        boolean storeResult = cfg.isStoreExpressionEvalResultsInContext();
+        Callable<Command> fn = new DelegateFn(expressionManager, ctx, a.getType(), a.getExpression(), a.getDefaultCommand(), storeResult);
 
         List<Timeout<Command>> timeouts = a.getTimeouts();
         if (timeouts != null && !timeouts.isEmpty()) {
@@ -85,7 +90,7 @@ public class ExpressionsReducer implements Reducer {
         // we apply new state of variables regardless of whether the call was
         // successful or not
         // TODO think about it
-        state = applyOutVariables(state, ctx, a.getOut());
+        state = VariablesHelper.applyOutVariables(expressionManager, state, ctx, a.getOut());
 
         return state;
     }
@@ -166,14 +171,16 @@ public class ExpressionsReducer implements Reducer {
         private final ExpressionType type;
         private final String expression;
         private final Command defaultCommand;
+        private final boolean storeResult;
 
-        public DelegateFn(ExpressionManager expressionManager, ExecutionContext ctx, ExpressionType type, String expression,
-                          Command defaultCommand) {
+        public DelegateFn(ExpressionManager expressionManager, ExecutionContext ctx, ExpressionType type,
+                          String expression, Command defaultCommand, boolean storeResult) {
             this.expressionManager = expressionManager;
             this.ctx = ctx;
             this.type = type;
             this.expression = expression;
             this.defaultCommand = defaultCommand;
+            this.storeResult = storeResult;
         }
 
         @Override
@@ -183,9 +190,17 @@ public class ExpressionsReducer implements Reducer {
             if (type == ExpressionType.DELEGATE) {
                 if (v instanceof JavaDelegate) {
                     ((JavaDelegate) v).execute(ctx);
+
+                    if (storeResult) {
+                        ctx.setVariable(ServiceTask.EXPRESSION_RESULT_VAR, null);
+                    }
                 } else {
                     throw new ExecutionException("Unexpected result type: " + v + ". Was expecting an instance of JavaDelegate");
                 }
+            }
+
+            if (storeResult) {
+                ctx.setVariable(ServiceTask.EXPRESSION_RESULT_VAR, v);
             }
 
             return defaultCommand;